fixes on PR review
