+++ code-recipes/README.md
@@ -6,3 +6,6 @@ Code recipes provides easy to read sample code for common use cases of Amazon Se
 | API | Use Case | Java                                                           | PHP                                                |
 | --- | -------- |----------------------------------------------------------------|----------------------------------------------------|
 | AWD | Inbound Shipment Creation | [Link](java/src/main/java/awd/InboundOrderCreationRecipe.java) | [Link](php/src/awd/InboundOrderCreationRecipe.php) |
+| A+ Content API | Submit an Image to A + Content | [Link](java/src/main/java/aplus/UploadImageForResouceRecipe.java) | - |
+| Messaging API | Submit an Invoice to Buyer | [Link](java/src/main/java/messaging/SendInvoiceToBuyerRecipe.java) | - |
+33
