+++ code-recipes/java/build.gradle
@@ -40,7 +40,7 @@ sourceSets {
 npmInstall.dependsOn("processTestResources")
 
 dependencies {
-	implementation 'software.amazon.spapi:spapi-sdk:1.5.0'
+	implementation 'software.amazon.spapi:spapi-sdk:1.3.0'
 	implementation 'com.squareup.okhttp3:okhttp:4.12.0'
 	implementation 'org.threeten:threetenbp:1.6.8'
 	testImplementation 'com.fasterxml.jackson.jr:jackson-jr-all:2.19.0'
