+++ code-recipes/java/src/main/java/util/Constants.java
@@ -1,5 +1,5 @@
 package util;
 
 public class Constants {
-    public static final String BACKEND_URL = "https://api.amazon.com";
+    public static final String BACKEND_URL = "http://localhost:3000";
 }
