+++ code-recipes/java/src/test/java/messaging/SendInvoiceToBuyerTest.java
@@ -1,0 +1,19 @@
+package messaging;
+
+import util.RecipeTest;
+
+import java.util.List;
+
+public class SendInvoiceToBuyerTest extends RecipeTest {
+
+    protected SendInvoiceToBuyerTest() {
+        super(
+                new SendInvoiceToBuyerRecipe(),
+                List.of(
+                    "messaging-getMessagingActionsForOrder",
+                    "messaging-createUploadDestination",
+                    ""
+                )
+        );
+    }
+}
