--- code-recipes/java/src/test/java/messaging/SendMessageToBuyerTest.java
@@ -1,24 +1,0 @@
-package messaging;
-
-import util.RecipeTest;
-
-import java.util.List;
-
-public class SendMessageToBuyerTest extends RecipeTest {
-
-    protected SendMessageToBuyerTest() {
-        super(
-                new SendInvoiceToBuyer(),
-                List.of(
-                    "",
-                    "",
-                    "messaging-getMessagingActionsForOrder",
-                    "",
-                    "",
-                    "messaging-createUploadDestination",
-                    "",
-                    ""
-                )
-        );
-    }
-}
